Moving the virtual field creation up to the module so we can get access to the injected configuration.
