+++ src/datasets/schemas/dataset.schema.ts
@@ -1,3 +1,4 @@
+import { ConfigService } from "@nestjs/config";
 import { Prop, Virtual, Schema, SchemaFactory } from "@nestjs/mongoose";
 import { ApiProperty, getSchemaPath } from "@nestjs/swagger";
 import { Document } from "mongoose";
@@ -269,44 +270,6 @@ export class DatasetClass extends OwnableClass {
   })
   @Prop({ type: String, required: true })
   version: string;
-
-  @ApiProperty({
-    type: [ExternalLinkClass],
-    required: false,
-    default: [],
-    description: "List of external links that involve this data set.",
-  })
-  @Virtual({
-    get: function (this: ExternalLinkClass[]) {
-      const thisDataSet = this;
-      const externalLinkDefinitions = [
-        {
-          "title": "Franzviewer",
-          "url_template": "https://franz.site.com/franzviewer?id=${dataset.pid}",
-          "description_template": "View ${dataset.numberOfFiles} files in the FranzViewer",
-          "filter": "(dataset.type == 'derived') && dataset.owner.includes('Franz')",
-        },
-        {
-          "title": "High Beam-Energy View",
-          "url_template": "https://beamviewer.beamline.net/highenergy?id=${dataset.pid}",
-          "description_template": "The high-energy beamviewer (value ${dataset.scientificMetadata?.beamEnergy?.value}) at beamCo",
-          "filter": "(dataset.scientificMetadata?.beamEnergy?.value > 20)",
-        }
-      ];
-      return externalLinkDefinitions.filter((d) => {
-        const filterFn = new Function("dataset", `return (${d.filter});`);
-        return filterFn(thisDataSet);
-      }).map((d) => {
-        const urlFn = new Function("dataset", `return (\`${d.url_template}\`);`);
-        const descriptionFn = new Function("dataset", `return (\`${d.description_template}\`);`);
-        return {  url: urlFn(thisDataSet),
-                  title: d.title,
-                  description: descriptionFn(thisDataSet)
-        };
-      });
-    },
-  })
-  externalLinks?: ExternalLinkClass[];
 
   @ApiProperty({
     type: "array",
